GITHUB_DC-Bot-commands/discord-bot/requirements.txt
discord.py
asyncio
FFmpeg
python-dotenv